FINALLY fixed the issue with reloading the page
I believe it had something to do with passing the posts as a prop and
then trying to use it asynchronously.  getPostID never had the posts
when it was trying to look up the PostID.

I’m still not 100% sure what the heck was going on here, but importing
the posts directly instead of passing it as a prop from the router,
worked.  I’ll probably figure it out later when I’m not as intensely
focused on this problem.  So relieved this is working now tho.